REVAI-3917: Split NlpModel to have separate enums for Translation and Summarization https://revinc.atlassian.net/browse/REVAI-3917
