+++ src/models/async/NlpModel.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-shadow */
-/** Supported model types for NLP tasks like summarization. */
+/** Supported model types for summarization. */
 export enum NlpModel {
     STANDARD = 'standard',
     PREMIUM = 'premium'
