+++ src/models/async/TranslationMode.ts
@@ -1,0 +1,6 @@
+/* eslint-disable no-shadow */
+/** Supported model types for translation. */
+export enum TranslationMode {
+    STANDARD = 'standard',
+    PREMIUM = 'premium'
+}
