+++ test/integration/test/async-translation-summarization.test.js
@@ -4,7 +4,8 @@ const { SummarizationJobStatus } = require('../../../dist/src/models/async/Summa
 const { SummarizationFormattingOptions } = require('../../../dist/src/models/async/SummarizationFormattingOptions');
 const { TranslationJobStatus } = require('../../../dist/src/models/async/TranslationJobStatus');
 const { JobStatus } = require('../../../dist/src/models/JobStatus');
-const { NlpModel } = require('../../../dist/src/models/NlpModel');
+const { NlpModel } = require('../../../dist/src/models/async/NlpModel');
+const { TranslationMode } = require('../../../dist/src/models/async/TranslationMode');
 
 test('async translation/summarization submit local file', async () => {
     const client = clientHelper.getAsyncClient();
@@ -23,7 +24,7 @@ test('async translation/summarization submit local file', async () => {
         target_languages : [
         {
             language:'es',
-            model:NlpModel.PREMIUM
+            model:TranslationMode.PREMIUM
         },
         {
             language:'ru'
@@ -60,7 +61,7 @@ test('async translation/summarization submit local file', async () => {
     expect(job.translation.completed_on).not.toBeNull();
     expect(job.translation.target_languages[0].status).toBe(TranslationJobStatus.Completed);
     expect(job.translation.target_languages[0].language).toBe("es");
-    expect(job.translation.target_languages[0].model).toBe(NlpModel.PREMIUM);
+    expect(job.translation.target_languages[0].model).toBe(TranslationMode.PREMIUM);
 
     expect(job.translation.target_languages[1].status).toBe(TranslationJobStatus.Completed);
     expect(job.translation.target_languages[1].language).toBe("ru");
@@ -109,7 +110,7 @@ test('async translation/summarization submit url', async () => {
         target_languages : [
         {
             language:'es',
-            model:NlpModel.PREMIUM
+            model:TranslationMode.PREMIUM
         },
         {
             language:'ru'
@@ -146,7 +147,7 @@ test('async translation/summarization submit url', async () => {
     expect(job.translation.completed_on).not.toBeNull();
     expect(job.translation.target_languages[0].status).toBe(TranslationJobStatus.Completed);
     expect(job.translation.target_languages[0].language).toBe("es");
-    expect(job.translation.target_languages[0].model).toBe(NlpModel.PREMIUM);
+    expect(job.translation.target_languages[0].model).toBe(TranslationMode.PREMIUM);
 
     expect(job.translation.target_languages[1].status).toBe(TranslationJobStatus.Completed);
     expect(job.translation.target_languages[1].language).toBe("ru");
